* commit '0c66bcfe2f6b6d1e22a3783d7b7985effc5d89e8':
  SELFB2C-2939 implementation hub assistance
  FIX DEEPLINKS TRANSFER OM & BUY PASS
  SELFB2C-2939 integration hub assistance
  Update icons dashboard et services hub
  Fix redirection rattachment on account page
  SELFB2C-2996 modal confirm delete numbers

+++ src/app/pages/rattached-phones-number/rattached-phones-number.page.ts
@@ -58,10 +58,18 @@ export class RattachedPhonesNumberPage implements OnInit {
     this.editable = !this.editable;
   }
 
-  deletePhoneNumbers(numero: string) {
-    const line = [numero];
-    this.accountService.deleteUserLinkedPhoneNumbers(line);
+  performAction(numero: string) {
+    if(this.editable) {
+      const line = [numero];
+      this.accountService.deleteUserLinkedPhoneNumbers(line);
+    } else {
+      this.switchPhoneNumber(numero);
+    }
   }
 
-  
+  switchPhoneNumber(msisdn) {
+    if (this.listRattachedNumbers.current === msisdn) return;
+    this.dashbServ.setCurrentPhoneNumber(msisdn);
+    this.goBack();
+  }
 }
